+++ illuin_inject/__init__.py
@@ -1,7 +1,6 @@
 from .annotated import annotated_arg
-from .bindings import AbstractModule, ClassBinding, FactoryBinding, InstanceBinding, ItemBinding, Module, \
-    MultiBinding, PrivateModule, SelfBinding
-from .factory import Factory
+from .bindings import AbstractModule, ClassBinding, InstanceBinding, ItemBinding, Module, MultiBinding, PrivateModule, \
+    ProviderBinding, SelfBinding
 from .injector import Injector
 from .provider import Provider
 from .scopes import ImmediateScope, PerLookupScope, SingletonScope, ThreadScope