+++ illuin_inject/bindings/multi_binding/item_binding.py
@@ -2,7 +2,7 @@ from typing import Generic, Type, Union
 
 import attr
 
-from illuin_inject.factory import Factory
+from illuin_inject.provider import Provider
 from illuin_inject.typings import EMPTY, InjectedT
 
 
@@ -10,4 +10,4 @@ from illuin_inject.typings import EMPTY, InjectedT
 class ItemBinding(Generic[InjectedT]):
     bound_type: Type[InjectedT] = EMPTY
     bound_instance: InjectedT = EMPTY
-    bound_factory: Union[Factory, Type[Factory]] = EMPTY
+    bound_provider: Union[Provider, Type[Provider]] = EMPTY