+++ illuin_inject/bindings/provider_binding/provider_binding_to_provider_adapter.py
@@ -1,33 +1,30 @@
 from illuin_inject.bindings.binding import Binding
 from illuin_inject.bindings.binding_to_provider_adapter import BindingToProviderAdapter
-from illuin_inject.bindings.instance_binding import FromInstanceProvider
 from illuin_inject.exceptions import NoBindingFound, NonInjectableTypeError
-from illuin_inject.factory import Factory
 from illuin_inject.injection_state import InjectionState
 from illuin_inject.provider import Provider
 from illuin_inject.target import Target
 from illuin_inject.typings import InjectedT
-from .factory_binding import FactoryBinding
-from .from_factory_provider import FromFactoryProvider
+from .from_provider_provider import FromProviderProvider
+from .provider_binding import ProviderBinding
 
 
 # pylint: disable=no-self-use, unused-argument
-class FactoryBindingToProviderAdapter(BindingToProviderAdapter[FactoryBinding]):
-    """Creates a Provider from an FactoryBinding."""
+class ProviderBindingToProviderAdapter(BindingToProviderAdapter[ProviderBinding]):
+    """Creates a Provider from a ProviderBinding."""
 
     def accept(self, binding: Binding[InjectedT], state: InjectionState) -> bool:
-        return isinstance(binding, FactoryBinding)
+        return isinstance(binding, ProviderBinding)
 
-    def create(self, binding: FactoryBinding[InjectedT], state: InjectionState) -> Provider[InjectedT]:
-        if isinstance(binding.bound_factory, Factory):
-            factory_provider = FromInstanceProvider(binding.bound_factory)
-        else:
-            factory_provider = state.provider_creator.get_provider(
-                Target(binding.bound_factory, binding.annotation),
-                state,
-            )
-        unscoped_provider = FromFactoryProvider(
-            factory_provider,
+    def create(self, binding: ProviderBinding[InjectedT], state: InjectionState) -> Provider[InjectedT]:
+        if isinstance(binding.bound_provider, Provider):
+            return binding.bound_provider
+        provider_provider = state.provider_creator.get_provider(
+            Target(binding.bound_provider, binding.annotation),
+            state,
+        )
+        unscoped_provider = FromProviderProvider(
+            provider_provider,
         )
         try:
             scope_provider = state.provider_creator.get_provider(Target(binding.scope), state)