+++ illuin_inject/factory.py
@@ -1,8 +1,0 @@
-from typing import Generic
-
-from .typings import InjectedT
-
-
-class Factory(Generic[InjectedT]):
-    def create(self) -> InjectedT:
-        raise NotImplementedError